requirements.txt
dj-database-url==0.5.0
Django>=2.2,<2.2.18
wagtail>=2.8.1
django-widget-tweaks==1.4.3
Markdown==3.0.1
python-decouple==3.1
django-simple-captcha==0.5.12
django_two_factor_auth==1.13
phonenumberslite==8.12.15